vicadmin version 0.1
Refactor vicadm.go into manageable chunks
Vicadmin now does setuid/setgid
Added ZIP file support
Fixed permissions of /var/log/vic files
Moving vicadmin page to template HTML file
Moving setuid/setgid from vicadmin into tether
