+++ cmd/vicadmin/html
@@ -1,0 +1,1 @@
+../../isos/vicadmin
