+++ cmd/vicadmin/vicadm.go
@@ -15,33 +15,25 @@
 package main
 
 import (
-	"archive/tar"
 	"bytes"
-	"compress/gzip"
-	"crypto/tls"
 	"errors"
 	"flag"
 	"fmt"
 	"io"
 	"io/ioutil"
-	"math"
-	"net"
 	"net/http"
 	"net/url"
 	"os"
 	"os/exec"
 	"os/signal"
 	"path"
-	"path/filepath"
 	"strings"
-	"sync"
 	"syscall"
 	"time"
 
 	"golang.org/x/net/context"
 
 	log "github.com/Sirupsen/logrus"
-	"github.com/docker/docker/pkg/tlsconfig"
 	"github.com/vmware/govmomi/object"
 	"github.com/vmware/govmomi/vim25/types"
 
@@ -87,6 +79,8 @@ var (
 		tls          bool
 	}
 
+	resources metadata.Resources
+
 	vchConfig metadata.VirtualContainerHostConfigSpec
 
 	defaultReaders map[string]entryReader
@@ -97,8 +91,20 @@ var (
 )
 
 func init() {
-	trace.Logger.Level = log.DebugLevel
 	defer trace.End(trace.Begin(""))
+	trace.Logger.Level = log.DebugLevel
+	_ = pprof.StartPprof("vicadmin", pprof.VicadminPort)
+
+	// We don't want to run this as root.
+	ud := syscall.Getuid()
+	gd := syscall.Getgid()
+	log.Info(fmt.Sprintf("Current UID/GID = %d/%d", ud, gd))
+	// TODO: Enable this after we figure out to NOT break the test suite with it.
+	// if ud == 0 {
+	// log.Errorf("Error: vicadmin must not run as root.")
+	// time.Sleep(60 * time.Second)
+	// os.Exit(1)
+	// }
 
 	flag.StringVar(&config.addr, "l", ":2378", "Listen address")
 	flag.StringVar(&config.dockerHost, "docker-host", "127.0.0.1:2376", "Docker host")
@@ -130,81 +136,6 @@ func init() {
 type Authenticator interface {
 	// Validate will validate a user and password combo and return a bool.
 	Validate(string, string) bool
-}
-
-func logFiles() []string {
-	defer trace.End(trace.Begin(""))
-
-	names := []string{}
-	for _, f := range logFileList {
-		names = append(names, fmt.Sprintf("%s/%s", logFileDir, f))
-	}
-
-	return names
-}
-
-func configureReaders() map[string]entryReader {
-	defer trace.End(trace.Begin(""))
-
-	pprofPaths := map[string]string{
-		// verbose
-		"verbose": "/debug/pprof/goroutine?debug=2",
-		// concise
-		"concise": "/debug/pprof/goroutine?debug=1",
-		"block":   "/debug/pprof/block?debug=1",
-		"heap":    "/debug/pprof/heap?debug=1",
-	}
-
-	pprofSources := map[string]string{
-		"docker":    pprof.GetPprofEndpoint(pprof.DockerPort).String(),
-		"portlayer": pprof.GetPprofEndpoint(pprof.PortlayerPort).String(),
-		"vicadm":    pprof.GetPprofEndpoint(pprof.VicadminPort).String(),
-		"vic-init":  pprof.GetPprofEndpoint(pprof.VCHInitPort).String(),
-	}
-
-	readers := map[string]entryReader{
-		"proc-mounts":  fileReader("/proc/mounts"),
-		"uptime":       commandReader("uptime"),
-		"df":           commandReader("df"),
-		"free":         commandReader("free"),
-		"netstat":      commandReader("netstat -ant"),
-		"iptables":     commandReader("sudo iptables -L -v"),
-		"iptables-nat": commandReader("sudo iptables -L -v -t nat"),
-		"ip-link":      commandReader("ip link"),
-		"ip-addr":      commandReader("ip addr"),
-		"ip-route":     commandReader("ip route"),
-		"lsmod":        commandReader("lsmod"),
-		// TODO: ls without shelling out
-		"disk-by-path":  commandReader("ls -l /dev/disk/by-path"),
-		"disk-by-label": commandReader("ls -l /dev/disk/by-label"),
-		// To check we are not leaking any fds
-		"proc-self-fd": commandReader("ls -l /proc/self/fd"),
-	}
-
-	// add the pprof collection
-	for sname, source := range pprofSources {
-		for pname, paths := range pprofPaths {
-			rname := fmt.Sprintf("%s/%s", sname, pname)
-			readers[rname] = urlReader(source + paths)
-		}
-	}
-
-	for _, path := range logFiles() {
-		// Strip off leading '/'
-		readers[path[1:]] = fileReader(path)
-	}
-
-	if config.vmPath == "" {
-		log.Info("vm-path not set, skipping datastore log collection")
-	} else {
-		err := findDatastore()
-
-		if err != nil {
-			log.Warning(err)
-		}
-	}
-
-	return readers
 }
 
 type entryReader interface {
@@ -436,87 +367,6 @@ func (r datastoreReader) open() (entry, error) {
 	return httpEntry(r.path, res)
 }
 
-type dlogReader struct {
-	c    *session.Session
-	name string
-	host *object.HostSystem
-}
-
-func findDiagnosticLogs(c *session.Session) (map[string]entryReader, error) {
-	defer trace.End(trace.Begin(""))
-
-	// When connected to VC, we collect vpxd.log and hostd.log for all cluster hosts attached to the datastore.
-	// When connected to ESX, we just collect hostd.log.
-	const (
-		vpxdKey  = "vpxd:vpxd.log"
-		hostdKey = "hostd"
-	)
-
-	logs := map[string]entryReader{}
-	var err error
-
-	if c.IsVC() {
-		logs[vpxdKey] = dlogReader{c, vpxdKey, nil}
-
-		var hosts []*object.HostSystem
-		if c.Cluster == nil && c.Host != nil {
-			hosts = []*object.HostSystem{c.Host}
-		} else {
-			hosts, err = c.Datastore.AttachedClusterHosts(context.TODO(), c.Cluster)
-			if err != nil {
-				return nil, err
-			}
-		}
-
-		for _, host := range hosts {
-			lname := fmt.Sprintf("%s/%s", hostdKey, host)
-			logs[lname] = dlogReader{c, hostdKey, host}
-		}
-	} else {
-		logs[hostdKey] = dlogReader{c, hostdKey, nil}
-	}
-
-	return logs, nil
-}
-
-func (r dlogReader) open() (entry, error) {
-	defer trace.End(trace.Begin(r.name))
-
-	name := r.name
-	if r.host != nil {
-		name = fmt.Sprintf("%s-%s", path.Base(r.host.InventoryPath), r.name)
-	}
-
-	m := object.NewDiagnosticManager(r.c.Vim25())
-	ctx := context.Background()
-
-	// Currently we collect the tail of diagnostic log files to avoid
-	// reading the entire file into memory or writing local disk.
-
-	// get LineEnd without any LineText
-	h, err := m.BrowseLog(ctx, r.host, r.name, math.MaxInt32, 0)
-	if err != nil {
-		return nil, err
-	}
-
-	// DiagnosticManager::DEFAULT_MAX_LINES_PER_BROWSE = 1000
-	start := h.LineEnd - 1000
-
-	h, err = m.BrowseLog(ctx, r.host, r.name, start, 0)
-	if err != nil {
-		return nil, err
-	}
-
-	var buf bytes.Buffer
-
-	for _, line := range h.LineText {
-		buf.WriteString(line)
-		buf.WriteString("\n")
-	}
-
-	return newBytesEntry(name+".log", buf.Bytes()), nil
-}
-
 func client() (*session.Session, error) {
 	defer trace.End(trace.Begin(""))
 
@@ -552,241 +402,6 @@ func findDatastore() error {
 	return nil
 }
 
-func tarEntries(readers map[string]entryReader, out io.Writer) error {
-	defer trace.End(trace.Begin(""))
-
-	r, w := io.Pipe()
-	t := tar.NewWriter(w)
-
-	wg := new(sync.WaitGroup)
-	wg.Add(1)
-
-	// stream tar to out
-	go func() {
-		_, err := io.Copy(out, r)
-		if err != nil {
-			log.Errorf("error copying tar: %s", err)
-		}
-		wg.Done()
-	}()
-
-	for name, r := range readers {
-		log.Infof("Collecting log with reader %s(%#v)", name, r)
-
-		e, err := r.open()
-		if err != nil {
-			log.Warningf("error reading %s(%s): %s\n", name, r, err)
-			continue
-		}
-
-		header := tar.Header{
-			Name:    name,
-			Size:    e.Size(),
-			Mode:    0640,
-			ModTime: time.Now(),
-		}
-
-		err = t.WriteHeader(&header)
-		if err != nil {
-			log.Errorf("Failed to write header for %s: %s", header.Name, err)
-			continue
-		}
-
-		log.Infof("%s has size %d", header.Name, header.Size)
-
-		// be explicit about the number of bytes to copy as the log files will likely
-		// be written to during this exercise
-		_, err = io.CopyN(t, e, e.Size())
-		_ = e.Close()
-		if err != nil {
-			log.Errorf("Failed to write content for %s: %s", header.Name, err)
-			continue
-		}
-	}
-
-	_ = t.Flush()
-	_ = w.Close()
-	wg.Wait()
-	_ = r.Close()
-
-	return nil
-}
-
-type server struct {
-	auth  Authenticator
-	l     net.Listener
-	addr  string
-	mux   *http.ServeMux
-	links []string
-}
-
-func (s *server) listen(useTLS bool) error {
-	defer trace.End(trace.Begin(""))
-
-	var err error
-
-	// Set options for TLS
-	tlsconfig := tlsconfig.ServerDefault
-	certificate, err := vchConfig.HostCertificate.Certificate()
-	if err != nil {
-		log.Errorf("Could not load certificate from config - running without TLS: %s", err)
-		// TODO: add static web page with the vic
-	} else {
-		tlsconfig.Certificates = []tls.Certificate{*certificate}
-	}
-
-	if !useTLS || err != nil {
-		s.l, err = net.Listen("tcp", s.addr)
-		return err
-	}
-
-	innerListener, err := net.Listen("tcp", s.addr)
-	if err != nil {
-		log.Fatal(err)
-		return err
-	}
-
-	s.l = tls.NewListener(innerListener, &tlsconfig)
-	return nil
-}
-
-func (s *server) listenPort() int {
-	return s.l.Addr().(*net.TCPAddr).Port
-}
-
-// handleFunc does preparatory work and then calls the HandleFunc method owned by the HTTP multiplexer
-func (s *server) handleFunc(link string, handler func(http.ResponseWriter, *http.Request)) {
-	defer trace.End(trace.Begin(""))
-
-	s.links = append(s.links, link)
-
-	if s.auth != nil {
-		authHandler := func(w http.ResponseWriter, r *http.Request) {
-			user, password, ok := r.BasicAuth()
-			if !ok || !s.auth.Validate(user, password) {
-				w.Header().Add("WWW-Authenticate", "Basic realm=vicadmin")
-				w.WriteHeader(http.StatusUnauthorized)
-				return
-			}
-			handler(w, r)
-		}
-		s.mux.HandleFunc(link, authHandler)
-		return
-	}
-
-	s.mux.HandleFunc(link, handler)
-}
-
-func (s *server) serve() error {
-	defer trace.End(trace.Begin(""))
-
-	s.mux = http.NewServeMux()
-
-	// tar of appliance system logs
-	s.handleFunc("/logs.tar.gz", s.tarDefaultLogs)
-
-	// tar of appliance system logs + container logs
-	s.handleFunc("/container-logs.tar.gz", s.tarContainerLogs)
-
-	// tail all logFiles
-	s.handleFunc("/logs/tail", func(w http.ResponseWriter, r *http.Request) {
-		s.tailFiles(w, r, logFiles())
-	})
-
-	for _, path := range logFiles() {
-		name := filepath.Base(path)
-		p := path
-
-		// get single log file (no tail)
-		s.handleFunc("/logs/"+name, func(w http.ResponseWriter, r *http.Request) {
-			http.ServeFile(w, r, p)
-		})
-
-		// get single log file (with tail)
-		s.handleFunc("/logs/tail/"+name, func(w http.ResponseWriter, r *http.Request) {
-			s.tailFiles(w, r, []string{p})
-		})
-	}
-
-	s.handleFunc("/", s.index)
-	server := &http.Server{
-		Handler: s.mux,
-	}
-
-	defaultReaders = configureReaders()
-
-	return server.Serve(s.l)
-}
-
-func (s *server) stop() error {
-	defer trace.End(trace.Begin(""))
-
-	if s.l != nil {
-		err := s.l.Close()
-		s.l = nil
-		return err
-	}
-
-	return nil
-}
-
-func (s *server) tarContainerLogs(res http.ResponseWriter, req *http.Request) {
-	defer trace.End(trace.Begin(""))
-
-	readers := defaultReaders
-
-	if config.Service != "" {
-		c, err := client()
-		if err != nil {
-			log.Errorf("failed to connect: %s", err)
-		} else {
-			// Note: we don't want to Logout() until tarEntries() completes below
-			defer c.Client.Logout(context.Background())
-
-			logs, err := findDatastoreLogs(c)
-			if err != nil {
-				log.Warningf("error searching datastore: %s", err)
-			} else {
-				for key, rdr := range logs {
-					readers[key] = rdr
-				}
-			}
-
-			logs, err = findDiagnosticLogs(c)
-			if err != nil {
-				log.Warningf("error collecting diagnostic logs: %s", err)
-			} else {
-				for key, rdr := range logs {
-					readers[key] = rdr
-				}
-			}
-		}
-	}
-
-	s.tarLogs(res, req, readers)
-}
-
-func (s *server) tarDefaultLogs(res http.ResponseWriter, req *http.Request) {
-	defer trace.End(trace.Begin(""))
-
-	s.tarLogs(res, req, defaultReaders)
-}
-
-func (s *server) tarLogs(res http.ResponseWriter, req *http.Request, readers map[string]entryReader) {
-	defer trace.End(trace.Begin(""))
-
-	res.Header().Set("Content-Type", "application/x-gzip")
-
-	z := gzip.NewWriter(res)
-
-	err := tarEntries(readers, z)
-	if err != nil {
-		log.Errorf("error taring logs: %s", err)
-	}
-
-	_ = z.Close()
-}
-
 type flushWriter struct {
 	f http.Flusher
 	w io.Writer
@@ -800,59 +415,18 @@ func (fw *flushWriter) Write(p []byte) (int, error) {
 	return n, err
 }
 
-func (s *server) tailFiles(res http.ResponseWriter, req *http.Request, names []string) {
-	defer trace.End(trace.Begin(""))
-
-	cc := res.(http.CloseNotifier).CloseNotify()
-
-	fw := &flushWriter{
-		f: res.(http.Flusher),
-		w: res,
-	}
-
-	// TODO: tail in Go, rather than shelling out
-	cmd := exec.Command("tail", append([]string{"-F"}, names...)...)
-	cmd.Stdout = fw
-	cmd.Stderr = fw
-
-	if err := cmd.Start(); err != nil {
-		log.Errorf("error tailing logs: %s", err)
-		return
-	}
-
-	defer cmd.Process.Kill()
-
-	go cmd.Wait()
-
-	<-cc
-}
-
-func (s *server) index(res http.ResponseWriter, req *http.Request) {
-	defer trace.End(trace.Begin(""))
-
-	fmt.Fprintln(res, "<html><head><title>VIC Admin</title></head><body><pre>")
-	for _, link := range s.links {
-		fmt.Fprintf(res, "<a href=\"%s\">%s</a><br/>\n", link, link)
-	}
-	fmt.Fprintln(res, "</pre></body></html>")
-}
-
 func main() {
 	if version.Show() {
 		fmt.Fprintf(os.Stdout, "%s\n", version.String())
 		return
 	}
 
-	defer trace.End(trace.Begin(""))
-	pprof.StartPprof("vicadmin", pprof.VicadminPort)
+	flag.Parse()
 
 	config.Service = vchConfig.Target.String()
 	config.ExtensionCert = vchConfig.ExtensionCert
 	config.ExtensionKey = vchConfig.ExtensionKey
 	config.ExtensionName = vchConfig.ExtensionName
-
-	log.Infof("VCH datastorePath: %+v", vchConfig.ImageStores)
-	log.Infof("From command-line: %+v", config.DatastorePath)
 
 	s := &server{
 		addr: config.addr,
