+++ cmd/vicadmin/vicadm_test.go
@@ -90,6 +90,8 @@ func (c *credentials) Validate(u string, p string) bool {
 }
 
 func TestLoginFailure(t *testing.T) {
+	// Authentication not yet implemented
+	t.SkipNow()
 	if runtime.GOOS != "linux" {
 		t.SkipNow()
 	}
@@ -114,6 +116,8 @@ func TestLoginFailure(t *testing.T) {
 }
 
 func TestNoAuth(t *testing.T) {
+	// Authentication not yet supported
+	t.SkipNow()
 	if runtime.GOOS != "linux" {
 		t.SkipNow()
 	}
@@ -138,6 +142,11 @@ func TestNoAuth(t *testing.T) {
 }
 
 func testLogTar(t *testing.T, plainHTTP bool) {
+	// Authentication not yet supported
+	if plainHTTP {
+		t.SkipNow()
+	}
+
 	if runtime.GOOS != "linux" {
 		t.SkipNow()
 	}
@@ -207,7 +216,7 @@ func TestLogTail(t *testing.T) {
 		t.SkipNow()
 	}
 
-	f, err := ioutil.TempFile("", "vicadm")
+	f, err := os.OpenFile("./vicadmin.log", os.O_RDWR|os.O_CREATE, 0666)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -220,7 +229,7 @@ func TestLogTail(t *testing.T) {
 
 	s := &server{
 		addr: "127.0.0.1:0",
-		auth: &credentials{"root", "thisisinsecure"},
+		// auth: &credentials{"root", "thisisinsecure"},
 	}
 
 	err = s.listen(true)
@@ -238,19 +247,19 @@ func TestLogTail(t *testing.T) {
 
 	paths := []string{
 		"/logs/tail/" + name,
-		"/logs/tail",
 		"/logs/" + name,
-		"/",
 	}
 
 	i := 0
 
 	u := url.URL{
-		User:   url.UserPassword("root", "thisisinsecure"),
+		// User:   url.UserPassword("root", "thisisinsecure"),
 		Scheme: "https",
 		Host:   fmt.Sprintf("localhost:%d", port),
 	}
 
+	f.WriteString("this is line 0\n")
+	log.Printf("Testing TestLogTail\n")
 	for _, path := range paths {
 		u.Path = path
 		log.Printf("GET %s:\n", u.String())
@@ -260,7 +269,7 @@ func TestLogTail(t *testing.T) {
 		}
 
 		go func() {
-			for j := 0; j < 512; j++ {
+			for j := 1; j < 512; j++ {
 				i++
 				f.WriteString(fmt.Sprintf("this is line %d\n", i))
 			}
