+++ isos/vicadmin/dashboard.html
@@ -1,0 +1,124 @@
+<!DOCTYPE html>
+<!-- saved from url=(0030)http://10.20.232.35/community/ -->
+<html itemscope="" itemtype="http://schema.org/Organization" lang="en-US">
+  <head>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+    <title>VCH Admin</title>
+    <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0">
+    <meta content="yes" name="apple-mobile-web-app-capable">
+    <meta name="apple-mobile-web-app-status-bar-style" content="black-translucent">
+    <!-- Styles -->
+    <link rel="stylesheet" href="css/clarity-ui.min.css">
+    <link rel="stylesheet" href="css/font-awesome.min.css">
+    <link rel="stylesheet" href="css/style.css">
+  </head>
+  <body id="body" class="page-community layout-community">
+    <div class="main-nav">
+      <header class="header">
+        <div class="branding">
+          <a href="http://www.vmware.com"><img src="images/vmw-logo.svg" class="logo"></a>
+          <span class="title">vSphere Integrated Containers</span>
+        </div>
+        <a role="button" class="btn btn-primary" href="http://www.github.com/vmware/vic">
+          <i class="fa fa-github"></i>
+          View on Github
+        </a>
+      </header>
+    </div>
+    <div class="main-container">
+      <main class="content-area">
+        <section class="community-top-section">
+          <h1 id="community">{{.Hostname}} Admin</h1>
+          <div class="row">
+            <div class="col-md-6">
+              <div class="card card-block">
+                <h3 class="card-title">Status</h3>
+                <ul class="status-card">
+                  <li>
+                    <span class="right">
+                      <i class="fa fa-check"></i>
+                    </span>
+                    Containers
+                  </li>
+                  <li>
+                    <span class="right warning">
+                      <i class="fa fa-exclamation-triangle"></i>
+                    </span>
+                    Network Connectivity
+                  </li>
+                  <li>
+                    <span class="right">
+                      <i class="fa fa-check"></i>
+                    </span>
+                    Firewall
+                  </li>
+                  <li>
+                    <span class="right">
+                      <i class="fa fa-check"></i>
+                    </span>
+                    Virtual Container Host (VCH)
+                  </li>
+                  <li>
+                    <span class="right">
+                      <i class="fa fa-check"></i>
+                    </span>
+                    License
+                  </li>
+                </ul>
+              </div>
+              <div class="card card-block">
+                <h3 class="card-title">Virtual Container Host Info</h3>
+                <p class="card-text">Docker Endpoint: 183.28.75.125:2376</p>
+                <p class="card-text">Storage: 420GB Remaining</p>
+              </div>
+            </div>
+            <div class="col-md-6">
+              <div class="card card-block">
+                <h3 class="card-title">Logs</h3>
+                <ul>
+                  <li><a href="/logs.zip">
+                      VCH Log Bundle
+                      </a></li>
+                  <li><a href="/container-logs.zip">
+                      VCH & Container Log Bundle
+                      </a></li>
+                  <li><a href="/logs/docker-personality.log">
+                      Docker Personality
+                      </a>
+                      <span class="right">
+                        <a href="/logs/tail/docker-personality.log">Live Log</a>
+                      </li>
+                  <li><a href="/logs/imagec.log">
+                      Docker Image Resolution
+                      </a>
+                      <span class="right">
+                        <a href="/logs/tail/imagec.log">Live Log</a>
+                      </li>
+                  <li><a href="/logs/port-layer.log">
+                      VCH Services (Low-Level)
+                      </a>
+                      <span class="right">
+                        <a href="/logs/tail/port-layer.log">Live Log</a>
+                      </li>
+                  <li><a href="/logs/init.log">
+                      VCH Initialization & Startup
+                      </a>
+                      <span class="right">
+                        <a href="/logs/tail/init.log">Live Log</a>
+                      </li>
+                  <li><a href="/logs/vicadmin.log">
+                      VIC Admin Server
+                      </a>
+                      <span class="right">
+                        <a href="/logs/tail/vicadmin.log">Live Log</a>
+                      </li>
+                </ul>
+              </div>
+            </div>
+          </div>
+        </section>
+      </main>
+    </div>
+  </body>
+</html>
